# .hacktogether/runs/87edeba4-9a80-4a73-858b-548fd9026da4/03-api-design/openapi.yaml
openapi: 3.1.0
info:
  title: Harvoost API
  version: 1.0.0
  description: |
    Harvoost — Harvest-style time-tracking SaaS for a single tenant on Azure.
    Provides endpoints for time entries, mood capture, scheduling, leave booking,
    two-stage approvals, exception handling, financial dashboards, Harvest-compatible
    Excel export, an RBAC-aware manager chatbot, and autonomous weekly summaries.

    ## Authentication
    Bearer token (opaque server-issued session id) in the `Authorization` header.
    The token is minted by `apps/api` after the Entra ID OIDC handshake completes.
    Token lifetime is 12h; clients refresh via `POST /v1/auth/refresh` (separate spec).

    ## Field casing
    All JSON field names, path parameters, and query parameters use `snake_case`.

    ## Error envelope
    Uniform shape across the entire API: `{ code, message, details? }`. See `ErrorResponse` schema.

    ## Pagination
    - **Cursor-based** for high-volume / append-mostly resources: time_entries, audit_log,
      exceptions, chatbot_conversations, chatbot_messages, leave_requests.
    - **Offset-based** for catalog-style resources: users, projects, clients, cost_rates, billable_rates.

    ## Idempotency
    `Idempotency-Key` header is REQUIRED on `POST /v1/time-entries/start`, `/stop`, and
    `/switch`. The server stores `(user_id, idempotency_key) → response` for 5 minutes;
    retried requests with the same key return the original response. Reuse of the same
...

    ## ISO 8601
...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/projects:
    get:
      tags: [Projects]
      summary: List projects (RBAC-scoped, offset-paginated)
      description: |
        Admin and FinMgr see all projects. Managers see projects in their anchored set.
        Employees see projects they are a member of.
      operationId: listProjects
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: client_id
          schema: { type: integer, format: int64 }
        - in: query
          name: billing_mode
          schema: { $ref: '#/components/schemas/BillingMode' }
        - in: query
          name: is_active
          schema: { type: boolean }
        - in: query
          name: search
          schema: { type: string }
      responses:
        '200':
          description: Paginated projects.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OffsetPaginationMeta'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Project' }
                      scope_meta: { $ref: '#/components/schemas/ScopeMeta' }
                    required: [data, scope_meta]
              example:
                page: 1
                page_size: 50
                total_count: 8
                data:
                  - id: 42
                    client_id: 7
                    code: ATL
                    name: Atlas
                    billing_mode: hourly
                    currency: ZAR
                    hours_budget: 320
                    department: Delivery
                    is_active: true
                    created_at: 2026-04-01T08:00:00+00:00
                    updated_at: 2026-05-10T13:00:00+00:00
                scope_meta:
                  visible_users: 12
                  visible_projects: 8
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Projects]
      summary: Create project (Admin only)
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProjectRequest' }
            example:
              client_id: 7
              code: ATL
              name: Atlas
              billing_mode: hourly
              currency: ZAR
              hours_budget: 320
              department: Delivery
      responses:
        '201':
          description: Project created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /v1/projects/{project_id}:
    parameters:
      - $ref: '#/components/parameters/ProjectIdPath'
    get:
      tags: [Projects]
      summary: Get project by id
      description: RBAC-scoped. Out-of-scope drill-down returns 403 RBAC_FORBIDDEN.
      operationId: getProject
      responses:
        '200':
          description: Project.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Projects]
      summary: Update project
      description: |
        Admin may update billing_mode, code, name, owner, is_active, hours_budget, department.
        FinMgr may update only the rate-related fields (billable_rate via project_billable_rates
        is a separate sub-resource — this endpoint only touches project-level scalar fields
        that FinMgr is allowed to manage: currency, hours_budget, fixed_fee_amount).
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProjectRequest' }
            example:
              name: Atlas v2
              hours_budget: 480
      responses:
        '200':
          description: Updated project.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /v1/projects/{project_id}/members:
    parameters:
      - $ref: '#/components/parameters/ProjectIdPath'
    get:
      tags: [Projects]
      summary: List members of a project (Admin only)
      description: |
        Admin-only. Lists the project's members with offset pagination. The caller must hold
        the `admin` role; other roles receive 403 `RBAC_FORBIDDEN`. Each row carries the
        optional `user_display_name`/`user_email` projections the admin members table renders.
      operationId: listProjectMembers
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Paginated project members.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OffsetPaginationMeta'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/ProjectMember' }
                    required: [data]
              example:
                page: 1
                page_size: 50
                total_count: 2
                data:
                  - id: 301
                    project_id: 42
                    user_id: 18
                    user_display_name: Bob Example
                    user_email: bob@harvoost.example.com
                    joined_at: 2026-05-22
                    left_at: null
                  - id: 302
                    project_id: 42
                    user_id: 19
                    user_display_name: Carol Example
                    user_email: carol@harvoost.example.com
                    joined_at: 2026-05-20
                    left_at: null
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Projects]
      summary: Add a member to a project (Admin only)
      operationId: addProjectMember
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddProjectMemberRequest' }
            example:
              user_id: 18
              joined_at: 2026-05-22
      responses:
        '201':
          description: Member added.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectMember' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/projects/{project_id}/members/{user_id}:
    parameters:
      - $ref: '#/components/parameters/ProjectIdPath'
      - $ref: '#/components/parameters/UserIdPath'
    delete:
      tags: [Projects]
      summary: Remove a member from a project (Admin only; soft delete sets left_at)
      operationId: removeProjectMember
      responses:
        '204':
          description: Member removed.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/projects/{project_id}/managers:
    parameters:
      - $ref: '#/components/parameters/ProjectIdPath'
    get:
      tags: [Projects]
      summary: List managers anchored to a project (Admin only)
      description: |
        Admin-only. Lists the manager anchors for this project with offset pagination. The
        caller must hold the `admin` role; other roles receive 403 `RBAC_FORBIDDEN`. Each row
        carries the optional `manager_display_name`/`manager_email` projections the admin
        managers table renders.
      operationId: listProjectManagers
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Paginated project manager anchors.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OffsetPaginationMeta'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/ProjectManagerAnchor' }
                    required: [data]
              example:
                page: 1
                page_size: 50
                total_count: 1
                data:
                  - id: 401
                    project_id: 42
                    manager_id: 17
                    manager_display_name: Alice Example
                    manager_email: alice@harvoost.example.com
                    assigned_at: 2026-05-01T08:00:00+00:00
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Projects]
      summary: Anchor a manager to this project (Admin only)
      operationId: addProjectManager
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddProjectManagerRequest' }
            example:
              manager_id: 17
      responses:
        '201':
          description: Manager anchored.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectManagerAnchor' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/projects/{project_id}/managers/{user_id}:
    parameters:
      - $ref: '#/components/parameters/ProjectIdPath'
      - $ref: '#/components/parameters/UserIdPath'
    delete:
      tags: [Projects]
      summary: Remove a manager anchor from this project (Admin only)
      operationId: removeProjectManager
      responses:
        '204':
          description: Anchor removed.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ============================================================
  # PROJECT TASKS
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Opaque server-issued session token (the API stores sessions in the `sessions` table
        — the token itself does not encode claims). The `bearerFormat: JWT` annotation here is
        a convention for OpenAPI tooling; the actual format is opaque (32+ bytes base64url).
        Issued by `POST /v1/auth/oidc/callback` after a successful Entra ID OIDC handshake.
  parameters:
...
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
      description: 1-based page index for offset-paginated collections.
    PageSize:
      in: query
      name: page_size
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
      description: Items per page (1..200, default 50).
...
    UserIdPath:
      in: path
      name: user_id
      required: true
      schema: { type: integer, format: int64 }
    ProjectIdPath:
      in: path
      name: project_id
      required: true
      schema: { type: integer, format: int64 }
...
  responses:
    BadRequest:
      description: Malformed request (e.g., bad JSON, bad date format).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            code: VALIDATION_FAILED
            message: Field `start_at` is not a valid ISO 8601 datetime with offset.
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            code: VALIDATION_FAILED
            message: Missing Authorization header.
    Forbidden:
      description: |
        Authenticated but not allowed (RBAC). Out-of-scope drill-downs return this code.
        Note: chatbot "no data accessible" responses are 200s with a polite text reply,
        NOT 403s.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            code: RBAC_FORBIDDEN
            message: You do not have access to this resource.
    NotFound:
      description: Resource not found (or hidden by RBAC; the API does not distinguish).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            code: NOT_FOUND
            message: Resource not found.
    ValidationFailed:
      description: Request body or query parameters failed validation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            code: VALIDATION_FAILED
            message: One or more fields failed validation.
            details:
              fields:
                - field: end_date
                  error: must be >= start_date
...
  schemas:
...
    BillingMode:
      type: string
      enum: [hourly, fixed_fee, non_billable]
...
    ErrorCode:
      type: string
      enum:
        - RBAC_FORBIDDEN
        - ENTRY_LOCKED
        - CHATBOT_DISABLED
        - IDEMPOTENCY_CONFLICT
        - VALIDATION_FAILED
        - NOT_FOUND
        - RATE_LIMITED
        - LLM_UNAVAILABLE
        - OIDC_FAILURE
        - K_ANONYMITY_THRESHOLD
      description: |
        Canonical machine-readable error codes returned in the `code` field of every error
        envelope. See API_NOTES.md for the semantic meaning of each code.

    # ---------- Error & pagination ----------
    ErrorResponse:
      type: object
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message:
          type: string
          description: Human-readable English explanation. Safe to display to end users.
        details:
          description: Optional structured context (e.g., `fields[]` for validation errors).
          oneOf:
            - { type: object, additionalProperties: true }
            - { type: array, items: { type: object, additionalProperties: true } }
      required: [code, message]
      examples:
        - code: ENTRY_LOCKED
          message: Cannot edit entry — currently in status manager_approved.
        - code: RBAC_FORBIDDEN
          message: You do not have access to this resource.
    ScopeMeta:
      type: object
      description: |
        Returned on every RBAC-scoped LIST response so clients can render meaningful
        empty states ("you have no projects in scope yet — contact your admin").
      properties:
        visible_users:
          type: integer
          description: Number of users in the requester's cascade scope (sentinel `-1` if Admin/FinMgr unrestricted).
        visible_projects:
          type: integer
      required: [visible_users, visible_projects]
...
    OffsetPaginationMeta:
      type: object
      description: Offset-based pagination metadata.
      properties:
        page: { type: integer }
        page_size: { type: integer }
        total_count: { type: integer }
      required: [page, page_size, total_count]

    # ---------- Auth ----------
...
    Project:
      type: object
      properties:
        id: { type: integer, format: int64 }
        client_id: { type: integer, format: int64 }
        code:
          oneOf:
            - { type: string }
            - { type: 'null' }
          description: Used in the Harvest `Project Code` export column.
        name: { type: string }
        billing_mode: { $ref: '#/components/schemas/BillingMode' }
        fixed_fee_amount:
          oneOf:
            - { type: number, format: float }
            - { type: 'null' }
        currency: { type: string, minLength: 3, maxLength: 3, description: ISO 4217. }
        hours_budget:
          oneOf:
            - { type: number, format: float }
            - { type: 'null' }
        department:
          oneOf:
            - { type: string }
            - { type: 'null' }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, client_id, name, billing_mode, currency, is_active, created_at, updated_at]
    CreateProjectRequest:
      type: object
      properties:
        client_id: { type: integer, format: int64 }
        code: { type: string, maxLength: 32 }
        name: { type: string, minLength: 1, maxLength: 200 }
        billing_mode: { $ref: '#/components/schemas/BillingMode' }
        fixed_fee_amount: { type: number, format: float }
        currency: { type: string, minLength: 3, maxLength: 3 }
        hours_budget: { type: number, format: float }
        department: { type: string }
      required: [client_id, name, billing_mode, currency]
    UpdateProjectRequest:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        billing_mode: { $ref: '#/components/schemas/BillingMode' }
        fixed_fee_amount: { type: number, format: float }
        currency: { type: string }
        hours_budget: { type: number, format: float }
        department: { type: string }
        is_active: { type: boolean }
      minProperties: 1
    AddProjectMemberRequest:
      type: object
      properties:
        user_id: { type: integer, format: int64 }
        joined_at: { type: string, format: date }
      required: [user_id]
    ProjectMember:
      type: object
      properties:
        id: { type: integer, format: int64 }
        project_id: { type: integer, format: int64 }
        user_id: { type: integer, format: int64 }
        user_display_name:
          oneOf:
            - { type: string }
            - { type: 'null' }
          description: Convenience join for the admin members table; may be absent on POST echoes.
        user_email:
          oneOf:
            - { type: string, format: email }
            - { type: 'null' }
          description: Convenience join for the admin members table; may be absent on POST echoes.
        joined_at: { type: string, format: date }
        left_at:
          oneOf:
            - { type: string, format: date }
            - { type: 'null' }
      required: [id, project_id, user_id, joined_at]
    AddProjectManagerRequest:
      type: object
      properties:
        manager_id: { type: integer, format: int64 }
      required: [manager_id]
    ProjectManagerAnchor:
      type: object
      properties:
        id: { type: integer, format: int64 }
        project_id: { type: integer, format: int64 }
        manager_id: { type: integer, format: int64 }
        manager_display_name:
          oneOf:
            - { type: string }
            - { type: 'null' }
          description: Convenience join for the admin managers table; may be absent on POST echoes.
        manager_email:
          oneOf:
            - { type: string, format: email }
            - { type: 'null' }
          description: Convenience join for the admin managers table; may be absent on POST echoes.
        assigned_at: { type: string, format: date-time }
      required: [id, project_id, manager_id, assigned_at]

    # ---------- Tasks ----------
...